requirements.txt
django==1.3.1
memcached
South==0.7.3
django-haystack
pysolr
